requirements.txt
Flask
Flask-Bcrypt
pymongo
python-dotenv